Change Google Chrome Apps "True" Death date?
Google Chrome Apps have been "supported and maintained" until now, but are being completely phased out by June 2020.
- https://en.wikipedia.org/wiki/Google_Chrome_App
- https://www.androidcentral.com/chrome-apps-are-going-away-starting-june-2020

So I updated the date close to March 2020, according to the wikipedia source, rather than the June 2020 of the androidcentral source.

Does Google Graveyard prefer to have the earlier phase-out date on Wikipedia / that was already used as the date close here? If so, we can just close this PR. But I thought at least submitting a more up-to-date death date for Chrome apps might be worth seeing if the core contributors of this repo think either date is more suitable.